OK-512: Siirretty harkinnanvaraisuuden tyypit yhteen paikkaan ja korjattu typo tiedoston nimessä
