+++ src/app/lib/types/harkinnanvaraiset-types.ts
@@ -1,4 +1,13 @@
-import { HarkinnanvarainenTila } from '../valintalaskenta-service';
+import { Hakemus } from './ataru-types';
+
+export type HarkinnanvarainenTila = 'HYVAKSYTTY' | 'EI_HYVAKSYTTY' | null;
+
+export type HarkinnanvaraisestiHyvaksytty = {
+  hakuOid: string;
+  hakukohdeOid: string;
+  hakemusOid: string;
+  harkinnanvaraisuusTila: HarkinnanvarainenTila;
+};
 
 export type HarkinnanvarainenTilaValue = HarkinnanvarainenTila | '';
 
@@ -6,3 +15,21 @@ export type HarkinnanvaraisetTilatByHakemusOids = Record<
   string,
   HarkinnanvarainenTilaValue
 >;
+
+export type HarkinnanvaraisuudenSyy =
+  | 'SURE_YKS_MAT_AI'
+  | 'SURE_EI_PAATTOTODISTUSTA'
+  | 'ATARU_YKS_MAT_AI'
+  | 'ATARU_ULKOMAILLA_OPISKELTU'
+  | 'ATARU_EI_PAATTOTODISTUSTA'
+  | 'ATARU_SOSIAALISET_SYYT'
+  | 'ATARU_OPPIMISVAIKEUDET'
+  | 'ATARU_KOULUTODISTUSTEN_VERTAILUVAIKEUDET'
+  | 'ATARU_RIITTAMATON_TUTKINTOKIELEN_TAITO'
+  | 'EI_HARKINNANVARAINEN'
+  | 'EI_HARKINNANVARAINEN_HAKUKOHDE';
+
+export type HakemuksenHarkinnanvaraisuus = Hakemus & {
+  harkinnanvaraisuudenSyy?: `harkinnanvaraisuuden-syy.${HarkinnanvaraisuudenSyy}`;
+  harkinnanvarainenTila: HarkinnanvarainenTila;
+};
