+++ src/app/lib/valintalaskentakoostepalvelu.ts
@@ -30,6 +30,7 @@ import {
 } from './valintaperusteet';
 import { ValintakoekutsutData } from './types/valintakoekutsut-types';
 import { HakutoiveValintakoeOsallistumiset } from './types/valintalaskentakoostepalvelu-types';
+import { HarkinnanvaraisuudenSyy } from './types/harkinnanvaraiset-types';
 
 const formSearchParamsForStartLaskenta = ({
   laskentaUrl,
@@ -497,20 +498,7 @@ export const putPistesyottoExcel = async ({
   return data;
 };
 
-export type HarkinnanvaraisuudenSyy =
-  | 'SURE_YKS_MAT_AI'
-  | 'SURE_EI_PAATTOTODISTUSTA'
-  | 'ATARU_YKS_MAT_AI'
-  | 'ATARU_ULKOMAILLA_OPISKELTU'
-  | 'ATARU_EI_PAATTOTODISTUSTA'
-  | 'ATARU_SOSIAALISET_SYYT'
-  | 'ATARU_OPPIMISVAIKEUDET'
-  | 'ATARU_KOULUTODISTUSTEN_VERTAILUVAIKEUDET'
-  | 'ATARU_RIITTAMATON_TUTKINTOKIELEN_TAITO'
-  | 'EI_HARKINNANVARAINEN'
-  | 'EI_HARKINNANVARAINEN_HAKUKOHDE';
-
-export type HakemuksenHarkinnanvaraisuustiedot = {
+type HakemuksenHarkinnanvaraisuustiedot = {
   hakemusOid: string;
   hakutoiveet: Array<{
     hakukohdeOid: string;
